* Refactor endpoints * Use same endpoint name convention * Fix returning HTTP OK/CREATED on unexciting peaks * Use Duration for completionTime * Use Long for number of minutes CompletionRequest * Move request parameters to RequestBody from query parameters * Remove unused config directory out of Application scope * Add CORS annotations 'just in case'

+++ src/main/java/pl/edu/agh/ki/io/api/PeakCompletionRequest.java
@@ -16,5 +16,5 @@ public class PeakCompletionRequest {
 
     @NotNull
     @SimpleSqlProtected
-    private int time;
+    private Long time;
 }
